Separate Entity and Dao from TopicDatabase.kt
